requirements.txt
# Core dependencies
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
undetected-chromedriver>=3.5.0
selenium>=4.15.0
alpaca-py>=0.21.0

# LLM APIs
openai>=1.3.0
requests>=2.31.0
tenacity>=8.2.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-selenium>=4.1.0
pytest-mock>=3.12.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
ruff>=0.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Logging and utilities
colorama>=0.4.0
rich>=13.7.0
